feat(ui): update color system to use light-dark oklch values
Implement light-dark color system using oklch values for improved
theme consistency. Removed obsolete hsl color definitions and refined
 media query to use breakpoint variables. Adjust pnpm-lock.yaml by
 removing deprecated notes for glob and inflight packages.
