+++ packages/ui/.storybook/with_theme_by_color_scheme.tsx
@@ -1,0 +1,28 @@
+import { DecoratorHelpers } from '@storybook/addon-themes'
+import { Decorator } from '@storybook/react'
+import { useEffect } from 'react'
+
+export function withThemeByColorScheme<Themes extends Record<string, string>>({
+	themes,
+	defaultTheme,
+}: {
+	themes: Themes
+	defaultTheme: keyof Themes
+}): Decorator {
+	DecoratorHelpers.initializeThemeState(Object.keys(themes), defaultTheme as string)
+	return (storyFn, context) => {
+		const { themeOverride } = DecoratorHelpers.useThemeParameters()
+		const selected = DecoratorHelpers.pluckThemeFromContext(context)
+
+		useEffect(() => {
+			const parentElement = document.querySelector('html')
+			const themeKey: keyof Themes | string = themeOverride || selected || defaultTheme
+
+			if (parentElement) {
+				parentElement.style.setProperty('color-scheme', String(themes[themeKey]))
+			}
+		}, [themeOverride, selected])
+
+		return storyFn()
+	}
+}
