+++ packages/ui/src/styles/styles.css
@@ -5,37 +5,37 @@
 @variant dark (&:is([data-theme="dark"] *));
 
 @theme {
-	/*--font-*: initial;*/
 	--font-comic: "Shantell Sans", cursive;
 	--font-comic--font-feature-settings: "cv11", "ss01";
 	--font-comic--font-variation-settings: "BNCE" 30, "INFM" 50, "SPAC" 0;
 
-	--color-border: var(--border);
-	--color-input: var(--input);
-	--color-ring: var(--ring);
-	--color-background: var(--background);
-	--color-foreground: var(--foreground);
+	--color-border: light-dark(oklch(0.922 0 89.876), oklch(0.269 0 89.876));
+	--color-input: light-dark(oklch(0.922 0 89.876), oklch(0.269 0 89.876));
+	--color-ring: light-dark(oklch(0.145 0 89.876), oklch(0.87 0 89.876));
+	--color-background: light-dark(oklch(1 0 89.876), oklch(0.145 0 89.876));
+	--color-foreground: light-dark(oklch(0.145 0 89.876), oklch(0.985 0 89.876));
 
-	--color-primary: var(--primary);
-	--color-primary-foreground: var(--primary-foreground);
+	--color-primary: light-dark(oklch(0.205 0 89.876), oklch(0.985 0 89.876));
+	--color-primary-foreground: light-dark(oklch(0.985 0 89.876), oklch(0.205 0 89.876));
 
-	--color-secondary: var(--secondary);
-	--color-secondary-foreground: var(--secondary-foreground);
+	--color-secondary: light-dark(oklch(0.97 0 89.876), oklch(0.269 0 89.876));
+	--color-secondary-foreground: light-dark(hsl(0 0% 9%), hsl(0 0% 98%));
 
-	--color-destructive: var(--destructive);
-	--color-destructive-foreground: var(--destructive-foreground);
+	--color-destructive: light-dark(oklch(0.637 0.208 25.331), oklch(0.396 0.133 25.723));
+	--color-destructive-foreground: light-dark(oklch(0.985 0 89.876), oklch(0.985 0 89.876));
 
-	--color-muted: var(--muted);
-	--color-muted-foreground: var(--muted-foreground);
+	--color-muted: light-dark(oklch(0.97 0 89.876), oklch(0.269 0 89.876));
+	--color-muted-foreground: light-dark(oklch(0.556 0 89.876), oklch(0.715 0 89.876));
 
-	--color-accent: var(--accent);
-	--color-accent-foreground: var(--accent-foreground);
+	--color-accent: light-dark(oklch(0.97 0 89.876), oklch(0.269 0 89.876));
+	--color-accent-foreground: light-dark(oklch(0.205 0 89.876), oklch(0.985 0 89.876));
 
-	--color-popover: var(--popover);
-	--color-popover-foreground: var(--popover-foreground);
+	--color-popover: light-dark(oklch(1 0 89.876), oklch(0.145 0 89.876));
+	--color-popover-foreground: light-dark(oklch(0.145 0 89.876), oklch(0.985 0 89.876));
 
-	--color-card: var(--card);
-	--color-card-foreground: var(--card-foreground);
+	--color-card: light-dark(oklch(1 0 89.876), oklch(0.145 0 89.876));
+	--color-card-foreground: light-dark(oklch(0.145 0 89.876), oklch(0.985 0 89.876));
+
 
 	--radius-lg: var(--radius);
 	--radius-md: calc(var(--radius) - 2px);
@@ -87,9 +87,9 @@
 @utility container {
 	margin-inline: auto;
 	padding-inline: 2rem;
-	@media (width >= theme(--breakpoint-sm)) {
-		max-width: none;
-	}
+	@media (width >= var(--breakpoint-sm)) {
+	max-width: none;
+}
 	@media (width >= 1400px) {
 		max-width: 1400px;
 	}
@@ -116,72 +116,16 @@
 @layer utilities {
 	@font-face {
 		font-family: "Shantell Sans";
+		font-style: normal;
+		font-weight: 400;
+		font-display: swap;
 		src: local(""), url("@fonts/ShantellSans.woff2") format("woff2");
 	}
 }
 
 @layer base {
 	:root {
-		--background: hsl(0 0% 100%);
-		--foreground: hsl(0 0% 3.9%);
-
-		--card: hsl(0 0% 100%);
-		--card-foreground: hsl(0 0% 3.9%);
-
-		--popover: hsl(0 0% 100%);
-		--popover-foreground: hsl(0 0% 3.9%);
-
-		--primary: hsl(0 0% 9%);
-		--primary-foreground: hsl(0 0% 98%);
-
-		--secondary: hsl(0 0% 96.1%);
-		--secondary-foreground: hsl(0 0% 9%);
-
-		--muted: hsl(0 0% 96.1%);
-		--muted-foreground: hsl(0 0% 45.1%);
-
-		--accent: hsl(0 0% 96.1%);
-		--accent-foreground: hsl(0 0% 9%);
-
-		--destructive: hsl(0 84.2% 60.2%);
-		--destructive-foreground: hsl(0 0% 98%);
-
-		--border: hsl(0 0% 89.8%);
-		--input: hsl(0 0% 89.8%);
-		--ring: hsl(0 0% 3.9%);
-
 		--radius: 0.75rem;
-	}
-
-	.dark,
-	:root[class~="dark"] {
-		--background: hsl(0 0% 3.9%);
-		--foreground: hsl(0 0% 98%);
-
-		--card: hsl(0 0% 3.9%);
-		--card-foreground: hsl(0 0% 98%);
-
-		--popover: hsl(0 0% 3.9%);
-		--popover-foreground: hsl(0 0% 98%);
-
-		--primary: hsl(0 0% 98%);
-		--primary-foreground: hsl(0 0% 9%);
-
-		--secondary: hsl(0 0% 14.9%);
-		--secondary-foreground: hsl(0 0% 98%);
-
-		--muted: hsl(0 0% 14.9%);
-		--muted-foreground: hsl(0 0% 63.9%);
-
-		--accent: hsl(0 0% 14.9%);
-		--accent-foreground: hsl(0 0% 98%);
-
-		--destructive: hsl(0 62.8% 30.6%);
-		--destructive-foreground: hsl(0 0% 98%);
-
-		--border: hsl(0 0% 14.9%);
-		--input: hsl(0 0% 14.9%);
-		--ring: hsl(0 0% 83.1%);
 	}
 
 	* {
